Added xml import and export

+++ src/main/kotlin/by/derovi/shapes/shapes/Sphere.kt
@@ -6,8 +6,8 @@ import kotlin.math.pow
 @Name("Сфера")
 class Sphere(
     @field:Name("Радиус")
-    val radius: Double = 0.0) : Shape() {
+    val radius: Int = 0) : Shape() {
 
     override val volume: Double
-        get() = 4.0 / 3.0 * Math.PI * radius.pow(3.0)
+        get() = 4.0 / 3.0 * Math.PI * radius.toDouble().pow(3.0)
 }
